Unset && clean tree
