+++ minishell/src/builtins/env.c
@@ -1,0 +1,22 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   builtin_env.c                                      :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: jmartin <jmartin@student.42lausanne.ch>    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/03/03 16:42:05 by jmartin           #+#    #+#             */
+/*   Updated: 2022/03/15 01:46:00 by jmartin          ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../../inc/minishell.h"
+
+void	env(t_shell *shell)
+{
+	int	i;
+
+	i = -1;
+	while (shell->envp[++i])
+		ft_printf("%s\n", shell->envp[i]);
+}
