+++ minishell/src/builtins/unset.c
@@ -1,44 +1,48 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   builtin_unset.c                                    :+:      :+:    :+:   */
+/*   unset.c                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: jmartin <jmartin@student.42lausanne.ch>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/14 12:05:05 by jmartin           #+#    #+#             */
-/*   Updated: 2022/03/14 17:47:25 by jmartin          ###   ########.fr       */
+/*   Updated: 2022/03/15 01:48:49 by jmartin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../inc/minishell.h"
+#include "../../inc/minishell.h"
 
-void	remove_envp(t_shell *shell, char *value, int size)
+void	*remove_envp(t_shell *shell, char *value, int size)
 {
-	int	i;
-	int	pos;
-	int	env_name_size;
+	int		i;
+	int		pos;
+	char	**ret;
+	int		env_name_size;
 
 	i = -1;
 	env_name_size = is_env_valid(value);
+	ret = malloc((size + 1) * sizeof(char *));
+	if (!ret)
+		return (NULL);
 	while (shell->envp[++i])
-	{
 		if (ft_strncmp(shell->envp[i], value, env_name_size) == 0)
-			shell->envp[i] = NULL;
-	}
+			pos = i;
 	i = -1;
-	while (shell->envp[++i])
+	while (++i < (size - 1))
 	{
-		if (shell->envp[i + 1] != NULL)
-		{
-			free(shell->envp[i]);
-			shell->envp[i] = ft_strdup(shell->envp[i + 1]);
-		}
+		if (i < pos)
+			ret[i] = ft_strdup(shell->envp[i]);
+		else
+			ret[i] = ft_strdup(shell->envp[i + 1]);
 	}
-	shell->envp[size - 1] = NULL;
+	ft_free_multi_tab(shell->envp);
+	shell->envp = ret;
+	return (ret);
 }
 
 void	unset(t_shell *shell)
 {
-	remove_envp(shell, shell->cmd.args[0], args_counter(shell->envp));
+	if (ft_strlen(shell->cmd.args[0]))
+		remove_envp(shell, shell->cmd.args[0], args_counter(shell->envp));
 	env(shell);
 }
