+++ minishell/src/main.c
@@ -6,7 +6,7 @@
 /*   By: jmartin <jmartin@student.42lausanne.ch>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/01 12:53:58 by jmartin           #+#    #+#             */
-/*   Updated: 2022/03/14 19:12:31 by jmartin          ###   ########.fr       */
+/*   Updated: 2022/03/14 23:55:26 by jmartin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
