+++ minishell/src/res/free.c
@@ -1,26 +1,16 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   utils.c                                            :+:      :+:    :+:   */
+/*   free.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: jmartin <jmartin@student.42lausanne.ch>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/03/07 10:52:05 by jmartin           #+#    #+#             */
-/*   Updated: 2022/03/14 17:39:10 by jmartin          ###   ########.fr       */
+/*   Created: 2022/03/15 01:50:22 by jmartin           #+#    #+#             */
+/*   Updated: 2022/03/15 01:52:24 by jmartin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../inc/minishell.h"
-
-int	args_counter(char **args)
-{
-	int	i;
-
-	i = -1;
-	while (args[++i])
-		;
-	return (i);
-}
+#include "../../inc/minishell.h"
 
 void	ft_free_tab(char *ptr)
 {
@@ -37,22 +27,4 @@ void	ft_free_multi_tab(char **ptr)
 		ft_free_tab(ptr[i]);
 	if (ptr)
 		free(ptr);
-}
-
-char	*append(char before, char *str, char after)
-{
-	size_t	len;
-
-	len = ft_strlen(str);
-	if (before)
-	{
-		ft_memmove(str + 1, str, ++len);
-		*str = before;
-	}
-	if (after)
-	{
-		str[len] = after;
-		str[len + 1] = 0;
-	}
-	return (str);
 }
